#11 水滴のアニメーションコードは水滴のComposableに移動する
