+++ app/src/main/kotlin/com/github/goutarouh/waterworksbureauclone/bottombar/Droplet.kt
@@ -1,10 +1,14 @@
 package com.github.goutarouh.waterworksbureauclone.bottombar
 
+import androidx.compose.animation.core.Animatable
+import androidx.compose.animation.core.TweenSpec
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.GenericShape
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Offset
@@ -13,6 +17,37 @@ import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.goutarouh.waterworksbureauclone.ui.theme.WaterworksBureauCloneTheme
+import kotlin.math.abs
+
+@Composable
+fun AnimationDroplet(
+    selectedIndex: Int,
+    preSelectedIndex: Int
+) {
+
+    val indicatorAngleAnimSpec = TweenSpec<Float>(durationMillis = 500)
+    val dropLetAngle = remember { Animatable(0f) }
+    LaunchedEffect(selectedIndex) {
+        when {
+            preSelectedIndex < selectedIndex -> dropLetAngle.animateTo(-20f, indicatorAngleAnimSpec)
+            preSelectedIndex > selectedIndex -> dropLetAngle.animateTo(20f, indicatorAngleAnimSpec)
+        }
+        dropLetAngle.animateTo(0f, indicatorAngleAnimSpec)
+    }
+
+
+    val dropLetAlpha = remember { Animatable(1f) }
+    LaunchedEffect(selectedIndex) {
+        if (preSelectedIndex == selectedIndex) return@LaunchedEffect
+        val diffPos = abs(preSelectedIndex - selectedIndex)
+        if (diffPos <= 1) return@LaunchedEffect
+        val durationMills = diffPos * 150
+        dropLetAlpha.animateTo(0f, TweenSpec(durationMillis = durationMills, delay = 200))
+        dropLetAlpha.animateTo(1f, TweenSpec(durationMillis = 0, delay = 200))
+    }
+
+    Droplet(angle = dropLetAngle.value, alpha = dropLetAlpha.value)
+}
 
 @Composable
 fun Droplet(
